+++ app/build.gradle
@@ -112,9 +112,10 @@ dependencies {
     testImplementation "androidx.room:room-testing:$room_version"
     testImplementation "androidx.paging:paging-common-ktx:$paging_version"
 
+    // Mockk is way better than mockito for kotlin (but only available for unit tests)
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'io.mockk:mockk:1.12.0'
     // Mockito
-    testImplementation 'org.mockito.kotlin:mockito-kotlin:3.2.0'
-    testImplementation 'org.mockito:mockito-inline:3.11.2'
     androidTestImplementation 'org.mockito:mockito-android:3.11.2'
 
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
