+++ app/src/androidTest/java/ch/epfl/reminday/ui/fragment/BirthdayListFragmentInstrumentedTest.kt
@@ -4,7 +4,7 @@ import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import ch.epfl.reminday.R
-import ch.epfl.reminday.SafeFragmentScenario
+import ch.epfl.reminday.testutils.SafeFragmentScenario
 import ch.epfl.reminday.data.birthday.BirthdayDao
 import ch.epfl.reminday.di.BirthdayDatabaseTestDI
 import ch.epfl.reminday.utils.Mocks
