+++ app/src/androidTest/java/ch/epfl/reminday/ui/view/DatePickerViewInstrumentedTest.kt
@@ -7,7 +7,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import ch.epfl.reminday.R
-import ch.epfl.reminday.SafeViewScenario
+import ch.epfl.reminday.testutils.SafeViewScenario
 import ch.epfl.reminday.testutils.NumberPickerTestUtils
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
